requirements.txt
altgraph==0.17.3
py-bcrypt==0.4
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
